WolfMusic/requirements.txt
Django==1.10.1
djangorestframework==3.4.6
pytaglib==1.3.0
